+++ src/cpp/CMakeLists.txt
@@ -29,9 +29,9 @@ set(${PROJECT_NAME}_source_files
     rtps/resources/ResourceEvent.cpp
     rtps/resources/TimedEvent.cpp
     rtps/resources/TimedEventImpl.cpp
-    rtps/resources/TimedCallback.cpp
     rtps/resources/AsyncWriterThread.cpp
     rtps/resources/AsyncInterestTree.cpp
+    rtps/timedevent/TimedCallback.cpp
     rtps/writer/RTPSWriter.cpp
     rtps/writer/StatefulWriter.cpp
     rtps/writer/ReaderProxy.cpp
